Split components folder, fixed update issue.
